+++ NEWS.md
@@ -5,7 +5,11 @@
 - Indenting with Tab in text area is now undo-able (with Ctrl+Z) (#19)
 - Added instructions to install from my R-universe in README (#13)
 - Import `utils`.
-- Fix encoding of reserved symbols (e.g. `#`) in URL queries for exporting notes as GitHub issues. 
+- Fix encoding of reserved symbols (e.g. `#`) in URL queries for exporting notes as GitHub issues.
+
+### API changes
+- Note names now can only contain alphanumeric characters, underscores (_) and 
+dots (.).
 
 ### Testing
 - Added `testthat` infrastructure.